Removed color from ID and fixed naming
