+++ packages/Tag/package.json
@@ -27,6 +27,7 @@
     "@igloo-ui/icon-button": "^0.2.11",
     "@igloo-ui/icons": "^1.3.0",
     "@igloo-ui/tokens": "^2.0.0",
+    "@igloo-ui/ellipsis": "^0.1.6",
     "classnames": "^2.3.1"
   },
   "browserslist": [
