+++ packages/Tag/src/components/color-icon.scss
