+++ packages/Tag/src/tag.scss
@@ -53,20 +53,20 @@
 
   /* Medium */
   --ids-tag-font-size: #{tokens.$font-size-3};
-  --ids-tag-height: 3.2rem;
+  --ids-tag-height: #{tokens.$space-5};
   --ids-tag-padding: #{tokens.$space-0} #{tokens.$space-2};
   --ids-tag-content-padding: #{tokens.$space-0} #{tokens.$space-2};
 
   /* Small */
   --ids-tag-font-size-small: #{tokens.$font-size-2};
-  --ids-tag-height-small: 2.4rem;
+  --ids-tag-height-small: #{tokens.$space-4};
   --ids-tag-padding-small: #{tokens.$space-0};
 
   /* XSmall */
   --ids-tag-height-xsmall: 2rem;
 
   /* Micro */
-  --ids-tag-height-micro: 1.6rem;
+  --ids-tag-height-micro: #{tokens.$space-3};
   --ids-tag-line-height-micro: #{tokens.$line-height-sm};
 
   /* === isError === */
